Add files via upload
E-Fixed Deposits Project Description
Overview
ABC Bank, a leading global bank with a presence in 56 countries, aims to transition to a fully digital banking model. As part of this initiative, the bank plans to introduce an E-Fixed Deposit (E-FD) feature on its existing online and mobile banking platforms. This feature will allow customers to create, renew, and close fixed deposits online, eliminating the need for physical branch visits.

The project involves capturing requirements, designing the feature, and implementing it for existing ABC Bank customers, with future plans to extend it to non-ABC Bank users.